+++ src/main/kotlin/us/frollo/frollosdk/messages/MessagesRx.kt
@@ -23,6 +23,7 @@ import us.frollo.frollosdk.extensions.sqlForMessages
 import us.frollo.frollosdk.mapping.toMessage
 import us.frollo.frollosdk.model.coredata.messages.ContentType
 import us.frollo.frollosdk.model.coredata.messages.Message
+import us.frollo.frollosdk.model.coredata.messages.MessageFilter
 
 /**
  * Fetch message by ID from the cache
@@ -36,25 +37,6 @@ fun Messages.fetchMessageRx(messageId: Long): Observable<Message?> {
 }
 
 /**
- * Fetch messages from the cache
- *
- * Fetches all messages if no params are passed.
- *
- * @param messageTypes List of message types to find matching Messages for (optional)
- * @param read Fetch only read/unread messages (optional)
- * @param contentType Filter the message by the type of content it contains (optional)
- *
- * @return Rx Observable object of List<Message> which can be observed using an Observer for future changes as well.
- */
-fun Messages.fetchMessagesRx(messageTypes: List<String>? = null, read: Boolean? = null, contentType: ContentType? = null): Observable<List<Message>> {
-    return db.messages().loadByQueryRx(sqlForMessages(messageTypes, read, contentType)).map { list ->
-        list.mapNotNull {
-            it.toMessage()
-        }
-    }
-}
-
-/**
  * Advanced method to fetch messages by SQL query from the cache
  *
  * @param query SimpleSQLiteQuery: Select query which fetches messages from the cache
@@ -65,6 +47,23 @@ fun Messages.fetchMessagesRx(messageTypes: List<String>? = null, read: Boolean? 
  */
 fun Messages.fetchMessagesRx(query: SimpleSQLiteQuery): Observable<List<Message>> {
     return db.messages().loadByQueryRx(query).map { list ->
+        list.mapNotNull {
+            it.toMessage()
+        }
+    }
+}
+
+/**
+ * Fetch messages from the cache with filters
+ *
+ * Fetches all messages if no params are passed.
+ *
+ * @param messageFilter [MessageFilter] object to apply filters
+ *
+ * @return Rx Observable object of List<Message> which can be observed using an Observer for future changes as well.
+ */
+fun Messages.fetchMessagesRx(messageFilter: MessageFilter = MessageFilter()): Observable<List<Message>> {
+    return db.messages().loadByQueryRx(sqlForMessages(messageFilter)).map { list ->
         list.mapNotNull {
             it.toMessage()
         }